initial stage of refactoring
