+++ tests/basic/src/test/java/org/jboss/hal/testsuite/test/configuration/messaging/server/ha/policy/SharedStoreColocatedTest.java
@@ -1,0 +1,8 @@
+package org.jboss.hal.testsuite.test.configuration.messaging.server.ha.policy;
+
+import org.jboss.arquillian.junit.Arquillian;
+import org.junit.runner.RunWith;
+
+@RunWith(Arquillian.class)
+public class SharedStoreColocatedTest extends AbstractHaPolicyTest {
+}
